Refactored `get_event_publisher`
