+++ llama_api/server/routers/v1.py
@@ -8,7 +8,6 @@ from dataclasses import dataclass, field
 from functools import partial
 from queue import Queue
 from random import choice
-from threading import Event
 from time import time
 from typing import (
     Any,
@@ -28,9 +27,7 @@ from anyio import (
     Semaphore,
     create_memory_object_stream,
     get_cancelled_exc_class,
-    move_on_after,
 )
-from anyio.streams.memory import MemoryObjectSendStream
 from fastapi import APIRouter, Request
 from fastapi.concurrency import iterate_in_threadpool, run_in_threadpool
 from orjson import OPT_INDENT_2, dumps
@@ -41,9 +38,7 @@ from llama_api.shared.config import MainCliArgs
 from ...mixins.completion import CompletionStatus
 from ...schemas.api import (
     ChatCompletion,
-    ChatCompletionChunk,
     Completion,
-    CompletionChunk,
     CreateChatCompletionRequest,
     CreateCompletionRequest,
     CreateEmbeddingRequest,
@@ -119,17 +114,29 @@ async def get_wix_with_semaphore(
 ) -> AsyncGenerator[int, None]:
     """Get the worker index (wix) for the key and acquire the semaphore"""
     global wix_metas
+
+    # Get the worker index (wix) with the lowest rank
+    # If the rank is -2, then the worker is processing the same model
+    # If the rank is -1, then the worker is not processing any model
+    # If the rank is greater than or equal to 0, then the worker is processing
+    # a different model
     worker_ranks = [
         get_worker_rank(wix_meta, request_key) for wix_meta in wix_metas
     ]
     min_rank = min(worker_ranks)
+
+    # Choose a random worker index (wix) with the lowest rank
     candidates = [i for i, rank in enumerate(worker_ranks) if rank == min_rank]
     if not candidates:
         raise LookupError("No available wix")
     wix_meta = wix_metas[choice(candidates)]
+
+    # Acquire the semaphore for the worker index (wix)
     async with wix_meta.semaphore:
+        # If client is already gone, then ignore the request
         if await request.is_disconnected():
             return
+        # Reserve the worker, it is now processing the request
         wix_meta.processed_key = request_key
         yield wix_meta.wix
 
@@ -143,60 +150,6 @@ def validate_item_type(item: Any, type: Type[T]) -> T:
         # The producer task has returned an invalid response
         raise TypeError(f"Expected type {type}, but got {type(item)} instead")
     return item
-
-
-def get_text_from_completion(
-    completion: Union[Completion, ChatCompletion]
-) -> str:
-    """Get the generated text from a completion"""
-    if "text" in completion["choices"][0]:
-        return completion["choices"][0]["text"]
-    return completion["choices"][0]["message"]["content"]
-
-
-def get_text_from_chunk(
-    chunk: Union[CompletionChunk, ChatCompletionChunk]
-) -> str:
-    """Get the generated text from a completion chunk"""
-    if "text" in chunk["choices"][0]:
-        return chunk["choices"][0]["text"]
-    return chunk["choices"][0]["delta"].get("content", "")
-
-
-async def get_event_publisher(
-    request: Request,
-    body: Union[
-        CreateChatCompletionRequest,
-        CreateCompletionRequest,
-    ],
-    inner_send_chan: MemoryObjectSendStream[bytes],
-    task: "Task[CompletionStatus]",
-    interrupt_signal: Event,
-    iterator: Iterator[Union[ChatCompletionChunk, CompletionChunk]],
-) -> None:
-    """Publish Server-Sent-Events (SSE) to the client"""
-    is_interrupted = False  # type: bool
-    async with inner_send_chan:
-        try:
-            async for chunk in iterate_in_threadpool(iterator):
-                await inner_send_chan.send(b"data: " + dumps(chunk) + b"\n\n")
-                if await request.is_disconnected():
-                    raise get_cancelled_exc_class()()
-            await inner_send_chan.send(b"data: [DONE]\n\n")
-        except get_cancelled_exc_class():
-            is_interrupted = True
-            with move_on_after(1, shield=True):
-                raise
-        finally:
-            # Cancel the producer task and set event,
-            # so the completion task can be stopped
-            interrupt_signal.set()
-            state = "Interrupted" if is_interrupted else "Completed"
-            try:
-                status = await wait_for(task, timeout=3)
-                log_request_and_response(body, status, state)
-            finally:
-                task.cancel()
 
 
 def get_streaming_iterator(
@@ -225,8 +178,11 @@ def log_request_and_response(
     state: Literal["Completed", "Interrupted"],
 ) -> None:
     """Log the request and response of the completion or embedding"""
+    # If the status is None, then the request has been interrupted
     if status is None:
         return
+
+    # Measure the elapsed time, and get information about the request
     elapsed_time = time() - status.started_at
     log_messages: List[str] = [f"elapsed time: {elapsed_time: .1f}s"]
     body_without_prompt = body.model_dump(
@@ -240,18 +196,20 @@ def log_request_and_response(
     if isinstance(status, EmbeddingStatus) and isinstance(
         body, CreateEmbeddingRequest
     ):
+        # Embedding usage is the number of characters in the input
+        # and the number of chunks in the embedding
         embed_usage = {
             "input_chars": len(body.input),
             "embedding_chunks": len(status.embedding["data"])
             if status.embedding
             else 0,
-        }
+        }  # type: Dict[str, int]
         log_messages.append(f"embedding chunks: {embed_usage}")
         embed_log = {
             "request": body_without_prompt,
             "input": body.input,
             "embedding": status.embedding,
-        }
+        }  # type: Dict[str, Any]
         logger.info(
             f"🦙 [{state} for {body.model}]: ({' | '.join(log_messages)})"
         )
@@ -264,6 +222,7 @@ def log_request_and_response(
     tokens_per_second = tokens / elapsed_time
     log_messages.append(f"tokens: {tokens}({tokens_per_second: .1f}tok/s)")
     if isinstance(body, CreateChatCompletionRequest):
+        # Log the chat completion status
         chat_log = {
             "request": body_without_prompt,
             "chat": [
@@ -276,15 +235,16 @@ def log_request_and_response(
                     "content": status.generated_text,
                 }
             ],
-        }
+        }  # type: Dict[str, Any]
     elif isinstance(body, CreateCompletionRequest):
+        # Log the text completion status
         chat_log = {
             "request": body_without_prompt,
             "prompt": {
                 "user": body.prompt,
                 "assistant": status.generated_text,
             },
-        }
+        }  # type: Dict[str, Any]
     else:
         return
     logger.info(f"🦙 [{state} for {body.model}]: ({' | '.join(log_messages)})")
@@ -316,22 +276,39 @@ async def create_chat_completion_or_completion(
         )
         if body.stream:
             send_chan, recv_chan = create_memory_object_stream(10)
+            chunk_iterator = get_streaming_iterator(
+                queue=queue,
+                first_response=validate_item_type(
+                    await run_in_threadpool(queue.get), type=dict
+                ),
+            )
+
+            async def get_event_publisher() -> None:
+                # Publish Server-Sent-Events (SSE) to the client
+                is_interrupted = False  # type: bool
+                send = send_chan.send
+                try:
+                    async for chunk in iterate_in_threadpool(chunk_iterator):
+                        await send(b"data: " + dumps(chunk) + b"\n\n")
+                    await send(b"data: [DONE]\n\n")
+                except get_cancelled_exc_class():
+                    is_interrupted = True
+                finally:
+                    send_chan.close()
+                    # Cancel the producer task and set event,
+                    # so the completion task can be stopped
+                    interrupt_signal.set()
+                    state = "Interrupted" if is_interrupted else "Completed"
+                    try:
+                        status = await wait_for(task, timeout=3)
+                        log_request_and_response(body, status, state)
+                    finally:
+                        task.cancel()
+
             return EventSourceResponse(
                 recv_chan,
-                data_sender_callable=partial(
-                    get_event_publisher,
-                    request=request,
-                    body=body,
-                    inner_send_chan=send_chan,
-                    task=task,
-                    interrupt_signal=interrupt_signal,
-                    iterator=get_streaming_iterator(  # type: ignore
-                        queue=queue,
-                        first_response=validate_item_type(
-                            await run_in_threadpool(queue.get), type=dict
-                        ),
-                    ),
-                ),
+                data_sender_callable=get_event_publisher,
+                ping=5,
             )
         else:
             # Cancel the producer task and set event,
